% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatOutput.R
\name{formatOutput}
\alias{formatOutput}
\title{formatOutput}
\usage{
formatOutput(rawOutput_dir, yearsToKeep)
}
\arguments{
\item{rawOutput_dir}{output directory containing the raw output from the MAGICC model}

\item{yearsToKeep}{vector of years to keep from MAGICC (which by default reports from 1750 on a yearly basis).}
}
\value{
a data.frame containing all the emissions scenarios' MAGICC7 warming pathways
}
\description{
formats the MAGICC7 raw output files
}
\examples{
  \dontrun{
    x <- formatOutput(rawOutput_dir)
  }
}
\author{
Michael Crawford
}
