% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMAGICC.R
\name{runMAGICC}
\alias{runMAGICC}
\title{runMAGICC}
\usage{
runMAGICC(emissions_fn, base_path)
}
\arguments{
\item{emissions_fn}{file name for the emissions (held in raw/input)}

\item{base_path}{root file path (the temporary directory)}
}
\description{
runs the MAGICC7 model on a specific emissions scenario
}
\examples{
  \dontrun{
    x <- runMAGICC(emissions_fn, base_path)
  }
}
\author{
Michael Crawford
}
