% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput.R
\name{writeInput}
\alias{writeInput}
\title{writeInput}
\usage{
writeInput(inputEmissions, rawInput_dir)
}
\arguments{
\item{inputEmissions}{data.frame representing the emissions to be written to disk and used by MAGICC}

\item{rawInput_dir}{directory in which to write the formatted .SCEN file (raw/input)}
}
\description{
writes the composite emissions scenario to a temporary folder, formatted to be read by MAGICC7
}
\examples{
  \dontrun{
    x <- formatInput(magpiemif_path, remindEmissions_path)
  }
}
\author{
Michael Crawford
}
