% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatInput.R
\name{formatInput}
\alias{formatInput}
\title{formatInput}
\usage{
formatInput(remindmif_path, magpiemif_path)
}
\arguments{
\item{remindmif_path}{file path for the reference REMIND emissions}

\item{magpiemif_path}{file path for the MAgPIE emissions}
}
\value{
a composite data.frame containing the scenario's MAgPIE emissions and reference REMIND emissions
}
\description{
converts MAgPIE emissions data, combining it with user-selected REMIND emissions, into the format
necessary to run MAGICC7
}
\examples{
  \dontrun{
    x <- formatInput(remindmif_path, magpiemif_path)
  }
}
\author{
Michael Crawford
}
